changed SketchRead to return a vector of uint16_t to fit with pile->data

+++ include/ram/minimizer_engine.hpp
@@ -47,7 +47,7 @@ class MinimizerEngine {
       float fraction,
       bool minhash = false);  // count k-mers in preconstructed minimizer index
 
-  std::vector<std::uint32_t> SketchRead(const std::unique_ptr<biosoup::NucleicAcid>& sequence, std::uint32_t step);
+  std::vector<std::uint16_t> SketchRead(const std::unique_ptr<biosoup::NucleicAcid>& sequence, std::uint32_t step);
   // set occurrence frequency threshold
   void Filter(double frequency);
