+++ src/minimizer_engine.cpp
@@ -644,12 +644,12 @@ std::vector<MinimizerEngine::Kmer> MinimizerEngine::Count(
   return dst;
 }
 
-std::vector<std::uint32_t> MinimizerEngine::SketchRead(
+std::vector<std::uint16_t> MinimizerEngine::SketchRead(
     const std::unique_ptr<biosoup::NucleicAcid>& sequence,
     std::uint32_t step){
 
   if (sequence->inflated_len < k_) {
-    return std::vector<std::uint32_t>{};
+    return std::vector<std::uint16_t>{};
   }
 
   std::uint64_t mask = (1ULL << (k_ * 2)) - 1;
@@ -685,7 +685,7 @@ std::vector<std::uint32_t> MinimizerEngine::SketchRead(
   std::uint64_t reverse_minimizer = 0;
   
 
-  std::vector<std::uint32_t> dst;
+  std::vector<std::uint16_t> dst;
 
   for (std::uint32_t i = 0; i < sequence->inflated_len; ++i) {
     std::uint64_t c = sequence->Code(i);
